+++ resources/js/app.js
@@ -1,1 +1,11 @@
 require('./bootstrap');
+
+import Alpine from 'alpinejs'
+
+window.Alpine = Alpine
+
+Alpine.start()
+
+
+require('@fortawesome/fontawesome-free/js/all.js');
+