+++ resources/views/errors/403.blade.php
@@ -2,6 +2,8 @@
 @section('title', 'Forbidden')
 	 
 @section('content')
-    <p class="fs-1">403</p>
+<div class="text-white mx-auto w-1/2">
+    <p class="text-8xl">403</p>
     <p>You do not have access to this page.</p>
+</div>
 @endsection