+++ resources/views/errors/503.blade.php
@@ -2,6 +2,8 @@
 @section('title', 'Maintenance')
 	 
 @section('content')
-    <p class="fs-1">Maintenance</p>
+<div class="text-white mx-auto w-1/2">
+    <p class="text-8xl">Maintenance</p>
     <p>Please check back later.</p>
+</div>
 @endsection